core_be/requirements.txt
asyncstdlib==3.10.4

boto3==1.21.42

celery==5.2.6
channels==3.0.4

dj-database-url==0.5.0
django-filter==2.3.0
django-graphql-jwt==0.3.4
django-storages==1.12.3
django-versatileimagefield==2.2
django-phonenumber-field==6.1.0
django-redis==4.12.1
django-prices==2.2.0
django-querycount==0.7.0
django==3.2.13
django-constance==2.8.0
django-picklefield==3.0.1
django-notifications-hq==1.7.0
django-otp==1.1.3

python-memcached==1.59

nested-lookup==0.2.23

graphene_django==2.15.0
#graphene_django==3.0.0b7
graphene-federation==0.1.0

firebase-admin==5.2.0
solana==0.23.3
psycopg2-binary==2.9.3

opentracing==2.4.0

phonenumbers==8.12.46
PyJWT==2.3.0

requests==2.27.1
redis==3.5.3

django-countries==7.3.2

dj-email-url==1.0.5

pytimeparse==1.1.8